Add files via upload
This part of the project uses a dataset of Overseas Filipino Workers (OFWs), specifically, the Statistical Profile of Registered OFWs (1981-2020) to analyze the relationship between educational attainment and migration patterns. The code performs data cleaning, reshaping, and aggregation to prepare the dataset for statistical modeling and visualization.

It includes exploratory data analysis, visualizations (line plots and stacked area charts), and regression modeling (Negative Binomial Regression) to investigate the impact of education levels and temporal effects on migration trends.